[mocks] Fixed YAML serialization of a stubbed object.
Previously, the mock proxy would be serialized when #to_yaml is called on a stubbed object.  The mock proxy includes a reference to a proc.  When the yaml string is deserialized with YAML.load, you'd get an "allocator undefined for Proc" TypeError.

This fix prevents stubbing from setting any new instance variables on the stubbed object, so that the YAMl serialization is not affected.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/793bd343ad1d0242e049fd7a3187c0f593498944.