[expectations] Fix matcher chaining to prevent name collisions.
Previously, #chain used #define_method on the class to define an instance method on all instances.  We want it to define the method on only this instance, so we need to delegate to the singleton class instead (which the recently added #define_method does for us!).

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/9c2971a8cbe4be660bf21b4078acfa0ec22ee10e.