[core] Output failures occurring in pending blocks
When using #pending with a block inside an example, it's nice to be able to
see what is still failing within that block. This is particularly useful for
an acceptance test, e.g. when using Capybara, because you may not want your CI
server to fail the build when committing work against an incomplete feature,
but you still want to see the relevant failures. This is analogous to using a
@wip tag with cucumber tests.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/a9b29ee3a3c8570b1ebfdf16ae37ada46e6c6fc5.