[expectations] Fix yield matchers to handle instance_eval.
It's debatable if instance_eval'ing a block can be considered yielding, but it's important that we use local variables in the lambdas rather than instance variables, because we don't want to modify the state of the receiver of #instance_eval.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/0e79594ba8635fbf709ce3a5358790bb261542f7.