[expectations] Raise an error if a yield matcher is used w/ an expect block that accepts no args.
The expect block must accept an arg, and pass it on to the method-under-test as a block, in order for these yield matchers to work properly.  Since this is atypical for an expect block, this error will help guide people when they use these matchers incorrectly.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/66a2f01372381581ccad91e79488fb3e881970cd.