[core] call #to_s on exception_message before calling #split
When 1.8.7 is shutting down its threads, and you're stuck in a Monitor,
it will eventually raise a SystemExit, and that SystemExit may have an
instance of Exception as its `#message` (instead of a String) which
causes the formatter to blow up.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/2d39c466664429bfa0f56f6cd4b1b419df8077bf.