[mocks] Replace duplicated README content with a link.
As per the conversation with @justinko and @dchelimsky:
https://github.com/rspec/rspec-mocks/pull/146#discussion_r919549

Note that this link is broken for now because this hasn't
yet been merged into master. But I figured it was better
not to use a working link just to the branch since that
branch will likely be deleted in the near future.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/0bd58502523550d3716400ec7aa92ffc4fb7f00d.