[core] Fix the way we autoload RSpec::Matchers.
`autoload`, besides being deprecated by Matz, does not work to require files
that are in unactivated gems--it only works with ruby's built in require, for
files that are available relative to a directory on the load path. Instead,
we use `const_missing` to make it work.

Closes rspec/rspec-core#647.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/e6ecc0b79601cbc6e2464ffc4950b32c69299996.