[mocks] Run all specs together and in isolation w/o bundler as part of the build.
We want these to always pass this way, so having it as part of the build is important.

This should help prevent a regression for rspec/rspec-mocks#174 and issues like it.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/16313fbe70651dc5d47f5ac9208ecc501de0f46e.