[expectations] Revert recent build change -- it's not working on Travis CI.
I need to head to bed in a bit and I don't want to leave the build
broken. I'll put this in a branch and play with it there to
get it to work.

- Revert "Improve how we find the path to the rspec exe."
  This reverts commit cafef45eceba3f38791bf9535d0fc52bf1829705.
- Revert "Run all specs together and in isolation for build."
  This reverts commit b0d5007173f8e0062fc21685a605c142e6d80a98.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/aa4530d40eab8fa3d453f4250cb5818d08bfe319.