[mocks] Remove `method_missing` inconsistency.
Based on conversations with @dchelimsky and others, we've decided it's best not
to raise an early error from `and_call_original` as that creates an
inconsistency since we sometimes don't know if `method_missing` will handle
the message or not.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/98ca8795dd38224f5300e3ec0207c8feee71a3a3.