[core] Updates BaseTextFormatter#colorize to check against a set of values
If it doesn't find the code in the set of values, it looks in the hash of VT_100 codes and values. If it can't find it there, it defaults to white.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/868a86144c6a822d8c25b7fac7b8fdd762eb8288.