[core] Refactor `subject` to be implemented in terms of `let`.
This is far simpler, and paves the way for supporting
`super` in both `let` and `subject`.

I had to rewrite a few of the subject specs; they were coupled
to ExampleGroup#subject returning a proc, but that was an
implementation detail of the old implementation.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/19dd6b2d4b25da0042f80bb8ff854672b913ba30.