[core] Re-instate the sleep -- the build failed on JRuby.
Locally, it passed on JRuby for me but failed on travis.
It failed due to reporting the time taken as 0% instead of 100%.

So, we need to ensure some time passes, but it doesn't need to be much.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/f256f1bdb3bc516ba622a8f34e810e70f31c0bb2.