[core] Ensure let block is not passed to super call.
Fix rspec/rspec-core#820. This is directly related to bug
http://bugs.ruby-lang.org/issues/8059.  It seems to affect Ruby 1.9.3
but not Ruby 2.0.0. As per the notes, it is a heisenbug, so simply
looking at it causes the bug to disappear.

Force the call to super to make sure the let block is not passed in.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/1e3bc17343fd921cf7d87b0214b51d2059c8ea12.