[mocks] Fix specs on 1.8.7.
Hashes are ordered on 1.9 but not on 1.8. Thus, we cannot count on
the order the mock proxies will be reset or verified since they are
held in a hash.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/554d80ce614ba366c76e420eb8b22d1c0770d96d.