[mocks] Allow the new syntax to be used w/o rspec-expectations.
Here we conditionally define `expect` in a superclass module
so that it is available, but `RSpec::Matchers` can still be
included later to override it.

Note: with this in place, `expect` won't be undefined when the
new syntax is disabled.  I think that's OK, though.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/ee9d82ee6110008d2712b50873e8b96269cfbb57.