[core] Fix brittle tests.
`#debug?` depends on whether or not the debugger is loaded. The tests for
it were brittle and could start to fail if you were running the specs with
the debugger loaded.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/0b6fd538edf28305373f0eacce5049e0b12f4d05.