[core] Fix brittle tests.
`#debug?` depends on whether or not the debugger is loaded. The tests for
it were brittle and could start to fail if you were running the specs with
the debugger loaded.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/b09c535d35f131011e74ac58f64acc2bf5896011.