[core] Move definition of implicit subject.
This is in prep for changing when the LetDefinitions module is included.
The use of `subject` from the `included` hook was causing problems when
I was trying to do that.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/bf2d58e4a3a11ec36e13ae995f384edc3e260c77.