[core] Move definition of implicit subject.
This is in prep for changing when the LetDefinitions module is included.
The use of `subject` from the `included` hook was causing problems when
I was trying to do that.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/b5f24314133f5edaec4d5dad3ccc1cb7e311761b.