[core] Move definition of implicit subject.
This is in prep for changing when the LetDefinitions module is included.
The use of `subject` from the `included` hook was causing problems when
I was trying to do that.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/583759a6a391fc2ea11fa273b9ca1b0f0a3694a7.