[core] Fixup the spec to work regardless of which solution we wind up with.
The original spec was focused on the exact fix that was originally put
in place, but we may wind up changing to an alternate fix that
would not be tested by the spec as it was.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/c4f010cc3651dedf218467bc7ee7816cc9560f51.