[mocks] Ensure the line in the warning is the block's source location.
Note: on 1.8.7, we can't always make that work because
Proc does not have `source_location`, and even if we
define `source_location` on it manually, it loses that
when we forward it on using `&block`.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/aab8dc9b3487dc51281078d2fa8c5f6a554b8245.