[expectations] Update caller filter.
- Include additional libs.
- Don't put it at `rspec/caller_filter`.
- Ignore the "same content as rspec-core" spec for now.

The way I had this was brittle and is causing problems in rspec-core.
In rspec-core I'm trying to update the regex and add some specs but
the fact that the file is at the same path in all 3 libs (which seemed
like a simple way to ensure only one of them loads) is causing
the _wrong_ version to be loaded in rspec-core's specs, so that it's not
picking up the changes I've made.  Putting the file at the same
path in all 3 libs was a terrible idea in retrospect :(.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/284066260560d6c9b358ea5731f6581adbb0f1b7.