[mocks] Pass __FILE__ and __LINE__ to `class_eval` properly.
This is important because w/o this, lines in the
eval'd code do not indicate what file or line
they are from when they are in a backtrace, and
that can cause CallerFilter to improperly consider
these lines to be non-rspec lines.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/1d0028b08f778dffc0beac2b877cb41bb3a424cc.