[core] Remove unnecessary module inclusion.
`SharedExampleGroup` needs to be extended, but
does not need to be included as well.

The specs had to be refactored a bit because they
declared shared examples in-line w/in an example,
which isn't a normal, supported use case.

Fixes rspec/rspec-core#1042.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/a3573aeb310729a41b191718ea3156507623412a.