[expectations] Use `include` instead of `extend`.
In RSpec 3 the body of `define` will be a
class definition and the body of `match` (etc)
will be a method def, so using `extend` to make
a method available to the `match` block will
not work.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/0ed732387a29ee50145a4ad87bec4ca591747c72.