[core] Revert "Swap output stream to current sandbox (Read why)."
This reverts commit a22d52467d936ff7bac9860b9d8275f9d0b55e43.

This is no longer needed with the new approach we are taking
to `output_stream` (e.g. having it default to $stdout), and
it was causing `command_line_spec.rb` to have a failure when
run in isolation.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/0ff9351fdc43bb80b2f09675856f3292966ebe48.