[mocks] Add specs for string vs symbol usage.
Each of these specs covers a case that was previously
uncovered. We use `to_sym` in a bunch of places to
normalize message names to a symbol. This is important,
but when I removed the `to_sym` calls, nothing was failing.

Without the `to_sym` calls, these specs fail, and they also
cause subsequent specs to fail because `reset` fails to
reset (since it winds up with two message doubles, `:foo`
and "foo", and resets the first and blows up on the second
since it already got reset -- that leaves the method double
around and it is attempted to be reset on reset for
subsequent specs).

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/8a7afc3de0a11b6d58df84de7b969cf38d1d1c9d.