[core] Rename deprecation printer classes.
For the new `raise_errors` option, we were using
the FilePrinter but not actually giving it a File
object. We use it because it immediately prints
deprecations, whereas the IOPrinter delays the printing
of errors until the end, so I think these names are
better.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/c96962a7bf7d5b747388febdb83c6c1b9f565037.