[core] Limit the methods of the `Hooks` module to just the public API.
Since this is mixed into a context where users write
code and define methods, we want to limit the number
of methods to just the ones that are part of the public
API.  The rest of the methods belong on the `HookCollections`
class better anyway.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/05195555b592d86c4846be2b7f1612af59c692a8.