[expectations] Post transpec cleanup.
Since we are testing the syntaxes here, we need to
be careful which ones we use in the tests. We can't
use `expect` when the syntax is configured to only
`:should`.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/197c95c81dbe0ccfda79329cd9196d624d1b35da.