[mocks] Tweak deprecation message a bit.
The deprecation is worded in general terms so it
seems odd for the replacement to be specific about
which message (particularly since it's not specific
about which object is passed to `allow`).

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/b4cc83adebd16208db75579a6a278bcd476f13e1.