[expectations] Fail match rather than error from `start_with` and `end_with`
…when the given object can't be indexed or is not ordered. We
want to be able to pass an alias like `a_collection_ending_with`
to a collection matcher (like `include`) which may contain
multiple types of objects, some of which can be indexed and are
ordered, and some not. We don't want an early `ArgumentError`
for these cases -- instead, it should just be treated as
not matching.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/c3f9f9b67901eb74e3e7718b954d74a551bd5839.