[expectations] Refactor: shrink the subproblems.
When recursing during the brute force depth-first
search, pass along only the indeterminate indexes,
so that it runs on a smaller subproblem. For this to
work, we had to switch to using a hash, so that it
can handle index gaps.

This provides mixed results: it makes some things faster,
some things slower. I think it's an improvement overall.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/e63799afaea5bb2ba150d0700025aa74cc58384e.