[expectations] Make DiffPresenter end all hash entries with comma
This fixes a bug where identical entries would show up as an added and a
subtracted line (that differed in the trailing comma) if they were the
last entry in one hash but not the other.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/0140b96c26737a533a2fb1deae5a4fb7039e8380.