[core] Standardize on `__send__` over `send`.
We've occasionally received bug reports related to
users having objects that redefine `send` (e.g.
`Email#send`), and using `__send__` is best to avoid
those problems.

Many of the receivers of `__send__` here are objects
we own and that do not (and never will) redefine `send`;
however, for consistency, it's good to standardize on
`__send__`. That way, we're less likely to forget it in
the future for a case where it does matter.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/bf2947d75e1330bbfabcb2fcf59c697d2a1fbe10.