[core] Remove skip-with-block feature.
This was initially intended as an replacement for `pending` with a block
(which has also been removed), but it's still really confusing.
Behaviour was the same as commenting out the code, which isn't useful.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/a3a01e43fced3a6ecee2d511fc3d436cb1dead67.