[expectations] Remove spec.
It doesn't really make sense to keep a spec that
specifies behavior when a specific gem is loaded
when there's nothing left in the project that is
custom code when that gem is loaded.

I thought it was worth committing the spec for the
history, though, so I kept it in the last commit.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/345201c7ba18e86069b0b23d4c53cbe468389fd3.