[mocks] Add some basic sanity check specs for `should` syntax.
Since we've migrated to the expect syntax we lost a lot
of test coverage for the should syntax.  For the most part
we don't need everything covered for both because it's the
same underlying mechanisms but some sanity tests are nice.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/126e6f146fe33ad2449694fd926ad863e8004b6d.