[expectations] Add missing require.
This file uses rspec-support’s MethodSignatureVerifier
but does not load it. It was working because rspec-mocks
loads it so it was loaded in the same process. It would
fail for users using rspec-expectations but not rspec-mocks,
though.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/9c2658ee6e58da79207de8d6ede1a660438f3897.