[expectations] Add missing require.
This file uses rspec-support’s MethodSignatureVerifier
but does not load it. It was working because rspec-mocks
loads it so it was loaded in the same process. It would
fail for users using rspec-expectations but not rspec-mocks,
though.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/5522ba60d8e5b0f645002262675d9b8e1bf2ae35.