[mocks] Fix edge case that triggered “can't add a new key into hash during iteration”.
This was triggered when a double was used as a `with` arg
but was otherwise unused — in this case, there was no mock
proxy for it, but while verifying one would get created
(due to internal implementation details). Doing so in a
`each_value` loop caused the error.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/a9d2100f85a22e752b4d8f668dbc80a7be80e646.