[expectations] Fix regression in `contain_exactly`.
The new implementation in 3.0.0 allowed this to pass:

expect([1]).to contain_exactly()

The problem was that the `actual_matches` hash was lazily
initialized in an `expected.each_with_index` loop. This
meant that if `expected` was empty, `actual_matches` would
be made empty as well, causing the match logic to pass.

Fixes rspec/rspec-expectations#581.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/b995e6fef458059d8622ecaa75a601f427241688.