[expectations] Extract evaluator classes for the nested vs sequential case.
This shortens the definition of `Compound` (allowing
us to drop the rubocop ClassLength setting to 90) and
groups all the logic for the nested case into a
single class so it’s not mixed up with the rest of
the compound logic.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/ce919e1d490811390c0525e33d6efcee0a595a85.