[expectations] Alias corrected constant name to old misspelling.
The misspelled version was released in 3.0. Removing
it in 3.x releases would be a SemVer violation, so we’ll
provide an alias.

This is a follow up to rspec/rspec-expectations#563.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/9a651ceb5d774fcc926046de22d91f1dfecf724d.