[core] Remove /gems/ from default backtrace exclusion patterns.
As discussed in rspec/rspec-core#1536 and rspec/rspec-core#1554, the failure can often be in a gem.

We may add an additional gem filtering API later on but for
now I just want to try this on a few projects and see how the
experience is.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/5dcde9e8e228223309c63eacc2fe69a17fe21a10.