[expectations] Fix all matcher exception when actual not enumerable
When the action value of passed to the `all` matcher is not enumerable,
rather than failing to match and returning an error it raises an
exception. This commit ensures that `actual` is enumerable before
proceeding with matching.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/5b4dc0ca6ad2ed11a7e5100854c7b4022010382c.