[core] Allow users to filter subdirs of their current dir.
This also greatly simplifies the backtrace formatter;
the system_exclusion_patterns approach made it more
complicated, as did always having an inclusion_pattern.

We only need the current directory as an inclusion_pattern
when it matches one of the built-in exclusion patterns.

Note that this was the approach that we originally had in rspec/rspec-core#843;
it got lost in the refactoring in rspec/rspec-core#1062 (af0b271c92ad2a3336eb70be7bf2c0de4325a568).
While it seemed like a simpler approach to always put the current dir
in the inclusion_patterns, in practice, this caused a sequence
of whack-a-mole regressions:

- rspec/rspec-core#1328 (fixed by rspec/rspec-core#1329)
- rspec/rspec-core#1604 (fixed by rspec/rspec-core#1616)
- rspec/rspec-core#1705 (fixed by this commit)

I'm hopeful that returning to only including the current dir
in `inclusion_patterns` if it matches a built-in exclusion pattern
will solve these issues once and for all.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/2cc12cefece83918b2e0737f43a72be52f195a16.