[support] Fix failing spec.
This spec was failing because of rspec/rspec-support#143 and rspec/rspec-support#145. Neither
PR failed when merged against the original master branch
but when both PRs were applied this began to fail because
of an interaction between the two.

Usually I would make a PR for this but I want to get our
builds back to green.

---
This commit was imported from https://github.com/rspec/rspec-support/commit/187845b1d46269b17edcf24a67744f98a8d1e994.