[core] Add missing integration spec coverage.
There are unit specs on the filter manager
that cover the filter manager’s involvement
in this behavior but nothing covered it as a whole,
and I’m working on a bug fix that may require
changes to the filter manager’s behavior, so
covering it with an integration spec ensures
we don’t regress.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/3cb6f39f32df096fa4bb9afe137ff6567dbfda77.